+ weapon_shareddefs.h
Shared definitions/enums/etc between all weapon code
