+++ mp/src/game/client/client_momentum.vpc
@@ -238,6 +238,7 @@ $Project "Client (Momentum)"
                 $File "$SRCDIR\game\shared\momentum\weapon\weapon_knife.h"
                 $File "$SRCDIR\game\shared\momentum\weapon\weapon_def.cpp"
                 $File "$SRCDIR\game\shared\momentum\weapon\weapon_def.h"
+                $File "$SRCDIR\game\shared\momentum\weapon\weapon_shareddefs.h"
             }
 
             $Folder "Util"
